Use homebrew core php installation
We will proceed by installing PHP 5.6, PHP 7.0, PHP 7.1 and PHP 7.2 and using a simple script to switch between them as we need. Up until the end of March 2018, all PHP related brews were handled by Homebrew/php tab, but that has been deprecated, so now we use what's available in the Homebrew/core package. This should be a better maintained, but is a much less complete, set of packages.

$ brew install php@5.6
$ brew install php@7.0
$ brew install php@7.1
$ brew install php@7.2